LLM_engineering_agentic_ai-main/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.75.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.4.0
Automat==25.4.16
babel==2.17.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.2
bleach==6.3.0
blinker==1.9.0
brotli==1.2.0
build==1.3.0
cachetools==6.2.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==1.3.5
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
constantly==23.10.4
contourpy==1.3.3
cryptography==46.0.3
cssselect==1.3.0
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
durationpy==0.10
email-validator==2.3.0
executing==2.2.1
fastapi==0.121.3
fastjsonschema==2.21.2
ffmpy==1.0.0
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.60.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.43.0
googleapis-common-protos==1.72.0
gradio==5.49.1
gradio_client==1.13.3
greenlet==3.2.4
griffe==1.15.0
groovy==0.1.2
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.1.5
humanfriendly==10.0
hyperlink==21.0.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
incremental==24.7.2
ipykernel==7.1.0
ipython==9.6.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
itemadapter==0.12.2
itemloaders==1.3.2
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.1
jmespath==1.0.1
joblib==1.5.2
json5==0.12.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.10
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.15
kiwisolver==1.4.9
kubernetes==34.1.0
langchain==1.1.0
langchain-anthropic==1.2.0
langchain-chroma==1.0.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.1.0
langchain-openai==1.1.0
langchain-text-splitters==1.0.0
langgraph==1.0.4
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.10
langsmith==0.4.49
lark==1.3.1
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.7
matplotlib-inline==0.2.1
mcp==1.23.1
mdurl==0.1.2
mistune==3.1.4
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.12.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.4.7
notebook_shim==0.2.4
numpy==2.3.5
oauthlib==3.3.1
ollama==0.6.1
onnxruntime==1.23.2
openai==2.8.1
openai-agents==0.6.1
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
ormsgpack==1.12.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parsel==1.10.0
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.5.0
plotly==6.5.0
posthog==5.4.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
Protego==0.5.0
protobuf==6.33.1
psutil==7.1.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
PyDispatcher==2.0.7
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.3.0
pyparsing==3.2.5
pypdf==6.4.0
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-http-client==3.3.7
python-json-logger==4.0.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
queuelib==1.8.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-file==3.0.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rpds-py==0.28.0
rsa==4.9.1
ruff==0.14.5
safehttpx==0.1.7
scikit-learn==1.7.2
scipy==1.16.3
Scrapy==2.13.3
selenium==4.38.0
semantic-version==2.10.0
Send2Trash==1.8.3
service-identity==24.2.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
some-package==0.1
sortedcontainers==2.4.0
soupsieve==2.8
SQLAlchemy==2.0.44
sse-starlette==3.0.3
stack-data==0.6.3
starlette==0.50.0
streamlit==1.51.0
sympy==1.14.0
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.6.0
tiktoken==0.12.0
tinycss2==1.4.0
tldextract==5.3.0
tokenizers==0.22.1
toml==0.10.2
tomlkit==0.13.3
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.32.0
trio-websocket==0.12.2
Twisted==25.5.0
typer==0.20.0
typer-slim==0.20.0
types-requests==2.32.4.20250913
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.38.0
uvloop==0.22.1
w3lib==2.3.1
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.2.14
webcolors==24.11.1
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wsproto==1.2.0
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zope.interface==8.0.1
zstandard==0.25.0
